entity design

+++ model/src/commonMain/kotlin/ponder/potato/model/game/math.kt
@@ -2,5 +2,5 @@ package ponder.potato.model.game
 
 import kotlin.math.pow
 
-fun factorValue(initialProgress: Double, level: Int, growthRate: Double) =
+fun factorValue(initialProgress: Int, level: Int, growthRate: Double) =
     initialProgress * growthRate.pow(level - 1)
